\name{ivaBSS-package}
\alias{ivaBSS-package}
\alias{ivaBSS}
\docType{package}
\title{
\packageTitle{ivaBSS}
}
\description{
\packageDescription{ivaBSS}
}
\details{

The package contains tools for independent vector analysis. The main functions to perform IVA are \code{"IVANewton"} and \code{"fastIVA"}. \code{"NewtonIVA"} performs Newton update based IVA and \code{"fastIVA"} performs fixed-point iteration based IVA. Both of the algorithms have multiple options for source density models.
}
\author{
Authors: Mika Sipilä, Klaus Nordhausen, Sara Taskinen

Maintainer: Mika Sipilä
}
\references{


\cite{Anderson, M., Adalı, T., & Li, X.-L. (2011). Joint blind source separation with multivariate
Gaussian model: Algorithms and performance analysis. IEEE Transactions on Signal
Processing, 60, 1672–1683. <doi:10.1109/TSP.2011.2181836>}

\cite{Anderson, M. (2013). Independent vector analysis: Theory, algorithms, and applications.
PhD dissertation, University of Maryland, Baltimore County.}
}

\keyword{ package }
